rule-agent/requirements.txt
langchain>=0.1.0
langgraph>=0.0.40
langchain-anthropic>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0
pytest>=7.0
pytest-cov>=4.0
python-dotenv>=1.0
python-docx>=1.1.0
pypdf>=4.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
